# Refactor: Pack WorldPoint into ints