--- a/src/main/java/shortestpath/pathfinder/CollisionMap.java
+++ b/src/main/java/shortestpath/pathfinder/CollisionMap.java
@@ -16,6 +16,7 @@ import net.runelite.api.coords.WorldPoint;
 import shortestpath.ShortestPathPlugin;
 import shortestpath.Transport;
 import shortestpath.Util;
+import shortestpath.WorldPointUtil;
 
 public class CollisionMap extends SplitFlagMap {
 
@@ -63,19 +64,22 @@ public class CollisionMap extends SplitFlagMap {
         return !n(x, y, z) && !s(x, y, z) && !e(x, y, z) && !w(x, y, z);
     }
 
-    private static WorldPoint pointFromOrdinal(WorldPoint start, OrdinalDirection direction) {
-        return new WorldPoint(start.getX() + direction.x, start.getY() + direction.y, start.getPlane());
+    private static int packedPointFromOrdinal(int startPacked, OrdinalDirection direction) {
+        final int x = WorldPointUtil.unpackWorldX(startPacked);
+        final int y = WorldPointUtil.unpackWorldY(startPacked);
+        final int plane = WorldPointUtil.unpackWorldPlane(startPacked);
+        return WorldPointUtil.packWorldPoint(x + direction.x, y + direction.y, plane);
     }
 
     public List<Node> getNeighbors(Node node, PathfinderConfig config) {
-        int x = node.position.getX();
-        int y = node.position.getY();
-        int z = node.position.getPlane();
+        final int x = WorldPointUtil.unpackWorldX(node.packedPosition);
+        final int y = WorldPointUtil.unpackWorldY(node.packedPosition);
+        final int z = WorldPointUtil.unpackWorldPlane(node.packedPosition);
 
         List<Node> neighbors = new ArrayList<>();
 
         @SuppressWarnings("unchecked") // Casting EMPTY_LIST to List<Transport> is safe here
-        List<Transport> transports = config.getTransports().getOrDefault(node.position, (List<Transport>)Collections.EMPTY_LIST);
+        List<Transport> transports = config.getTransportsPacked().getOrDefault(node.packedPosition, (List<Transport>)Collections.EMPTY_LIST);
 
         // Transports are pre-filtered by PathfinderConfig.refreshTransportData
         // Thus any transports in the list are guaranteed to be valid per the user's settings
@@ -112,12 +116,12 @@ public class CollisionMap extends SplitFlagMap {
 
         for (int i = 0; i < traversable.length; i++) {
             OrdinalDirection d = ORDINAL_VALUES[i];
-            WorldPoint neighbor = pointFromOrdinal(node.position, d);
+            int neighborPacked = packedPointFromOrdinal(node.packedPosition, d);
             if (traversable[i]) {
-                neighbors.add(new Node(neighbor, node));
+                neighbors.add(new Node(neighborPacked, node));
             } else if (Math.abs(d.x + d.y) == 1 && isBlocked(x + d.x, y + d.y, z)) {
                 @SuppressWarnings("unchecked") // Casting EMPTY_LIST to List<Transport> is safe here
-                List<Transport> neighborTransports = config.getTransports().getOrDefault(neighbor, (List<Transport>)Collections.EMPTY_LIST);
+                List<Transport> neighborTransports = config.getTransportsPacked().getOrDefault(neighborPacked, (List<Transport>)Collections.EMPTY_LIST);
                 for (int t = 0; t < neighborTransports.size(); ++t) {
                     Transport transport = neighborTransports.get(t);
                     neighbors.add(new Node(transport.getOrigin(), node));