--- a/src/main/java/shortestpath/pathfinder/CollisionMap.java
+++ b/src/main/java/shortestpath/pathfinder/CollisionMap.java
@@ -1,30 +1,25 @@
 package shortestpath.pathfinder;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import shortestpath.ShortestPathPlugin;
 import shortestpath.Transport;
-import shortestpath.Util;
 import shortestpath.WorldPointUtil;
 
-public class CollisionMap extends SplitFlagMap {
+public class CollisionMap {
 
     // Enum.values() makes copies every time which hurts performance in the hotpath
     private static final OrdinalDirection[] ORDINAL_VALUES = OrdinalDirection.values();
-    private static RegionExtent regionExtents = new RegionExtent(0, 0, 0, 0);
 
-    public CollisionMap(Map<Integer, byte[]> compressedRegions) {
-        super(compressedRegions, 2);
+    private final SplitFlagMap collisionData;
+
+    public CollisionMap(SplitFlagMap collisionData) {
+        this.collisionData = collisionData;
+    }
+
+    public boolean get(int x, int y, int z, int flag) {
+        return collisionData.get(x, y, z, flag);
     }
 
     public boolean n(int x, int y, int z) {
@@ -70,12 +65,16 @@ public class CollisionMap extends SplitFlagMap {
         return WorldPointUtil.packWorldPoint(x + direction.x, y + direction.y, plane);
     }
 
+    // This is only safe if pathfinding is single-threaded
+    private final List<Node> neighbors = new ArrayList<>(16);
+    private final boolean[] traversable = new boolean[8];
+
     public List<Node> getNeighbors(Node node, PathfinderConfig config) {
         final int x = WorldPointUtil.unpackWorldX(node.packedPosition);
         final int y = WorldPointUtil.unpackWorldY(node.packedPosition);
         final int z = WorldPointUtil.unpackWorldPlane(node.packedPosition);
 
-        List<Node> neighbors = new ArrayList<>();
+        neighbors.clear();
 
         @SuppressWarnings("unchecked") // Casting EMPTY_LIST to List<Transport> is safe here
         List<Transport> transports = config.getTransportsPacked().getOrDefault(node.packedPosition, (List<Transport>)Collections.EMPTY_LIST);
@@ -87,7 +86,6 @@ public class CollisionMap extends SplitFlagMap {
             neighbors.add(new TransportNode(transport.getDestination(), node, transport.getWait()));
         }
 
-        boolean[] traversable;
         if (isBlocked(x, y, z)) {
             boolean westBlocked = isBlocked(x - 1, y, z);
             boolean eastBlocked = isBlocked(x + 1, y, z);
@@ -97,20 +95,23 @@ public class CollisionMap extends SplitFlagMap {
             boolean southEastBlocked = isBlocked(x + 1, y - 1, z);
             boolean northWestBlocked = isBlocked(x - 1, y + 1, z);
             boolean northEastBlocked = isBlocked(x + 1, y + 1, z);
-            traversable = new boolean[] {
-                !westBlocked,
-                !eastBlocked,
-                !southBlocked,
-                !northBlocked,
-                !southWestBlocked && !westBlocked && !southBlocked,
-                !southEastBlocked && !eastBlocked && !southBlocked,
-                !northWestBlocked && !westBlocked && !northBlocked,
-                !northEastBlocked && !eastBlocked && !northBlocked
-            };
+            traversable[0] = !westBlocked;
+            traversable[1] = !eastBlocked;
+            traversable[2] = !southBlocked;
+            traversable[3] = !northBlocked;
+            traversable[4] = !southWestBlocked && !westBlocked && !southBlocked;
+            traversable[5] = !southEastBlocked && !eastBlocked && !southBlocked;
+            traversable[6] = !northWestBlocked && !westBlocked && !northBlocked;
+            traversable[7] = !northEastBlocked && !eastBlocked && !northBlocked;
         } else {
-            traversable = new boolean[] {
-                w(x, y, z), e(x, y, z), s(x, y, z), n(x, y, z), sw(x, y, z), se(x, y, z), nw(x, y, z), ne(x, y, z)
-            };
+            traversable[0] = w(x, y, z);
+            traversable[1] = e(x, y, z);
+            traversable[2] = s(x, y, z);
+            traversable[3] = n(x, y, z);
+            traversable[4] = sw(x, y, z);
+            traversable[5] = se(x, y, z);
+            traversable[6] = nw(x, y, z);
+            traversable[7] = ne(x, y, z);
         }
 
         for (int i = 0; i < traversable.length; i++) {
@@ -129,54 +130,5 @@ public class CollisionMap extends SplitFlagMap {
         }
 
         return neighbors;
-    }
-
-    public static CollisionMap fromResources() {
-        Map<Integer, byte[]> compressedRegions = new HashMap<>();
-        try (ZipInputStream in = new ZipInputStream(ShortestPathPlugin.class.getResourceAsStream("/collision-map.zip"))) {
-            int minX = Integer.MAX_VALUE;
-            int minY = Integer.MAX_VALUE;
-            int maxX = 0;
-            int maxY = 0;
-
-            ZipEntry entry;
-            while ((entry = in.getNextEntry()) != null) {
-                String[] n = entry.getName().split("_");
-                final int x = Integer.parseInt(n[0]);
-                final int y = Integer.parseInt(n[1]);
-                minX = Math.min(minX, x);
-                minY = Math.min(minY, y);
-                maxX = Math.max(maxX, x);
-                maxY = Math.max(maxY, y);
-
-                compressedRegions.put(
-                        SplitFlagMap.packPosition(x, y),
-                        Util.readAllBytes(in)
-                );
-            }
-
-            regionExtents = new RegionExtent(minX, minY, maxX, maxY);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-        return new CollisionMap(compressedRegions);
-    }
-
-    public static RegionExtent getRegionExtents() {
-        return regionExtents;
-    }
-
-    @RequiredArgsConstructor
-    @Getter
-    public static class RegionExtent {
-        public final int minX, minY, maxX, maxY;
-
-        public int getWidth() {
-            return maxX - minX;
-        }
-
-        public int getHeight() {
-            return maxY - minY;
-        }
     }
 }