--- a/src/main/java/shortestpath/pathfinder/Node.java
+++ b/src/main/java/shortestpath/pathfinder/Node.java
@@ -4,20 +4,31 @@ import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 import net.runelite.api.coords.WorldPoint;
+import shortestpath.WorldPointUtil;
 
 public class Node {
-    public final WorldPoint position;
+    public final int packedPosition;
     public final Node previous;
     public final int cost;
 
     public Node(WorldPoint position, Node previous, int wait) {
-        this.position = position;
+        this.packedPosition = WorldPointUtil.packWorldPoint(position);
         this.previous = previous;
-        this.cost = cost(previous, position, wait);
+        this.cost = cost(previous, wait);
     }
 
     public Node(WorldPoint position, Node previous) {
         this(position, previous, 0);
+    }
+
+    public Node(int packedPosition, Node previous, int wait) {
+        this.packedPosition = packedPosition;
+        this.previous = previous;
+        this.cost = cost(previous, wait);
+    }
+
+    public Node(int packedPosition, Node previous) {
+        this(packedPosition, previous, 0);
     }
 
     public List<WorldPoint> getPath() {
@@ -25,44 +36,41 @@ public class Node {
         Node node = this;
 
         while (node != null) {
-            path.add(0, node.position);
+            WorldPoint position = WorldPointUtil.unpackWorldPoint(node.packedPosition);
+            path.add(0, position);
             node = node.previous;
         }
 
         return new ArrayList<>(path);
     }
 
-    private static int cost(Node previous, WorldPoint current, int wait) {
+    public List<Integer> getPathPacked() {
+        List<Integer> path = new LinkedList<>();
+        Node node = this;
+
+        while (node != null) {
+            path.add(0, node.packedPosition);
+            node = node.previous;
+        }
+
+        return new ArrayList<>(path);
+    }
+
+    private int cost(Node previous, int wait) {
         int previousCost = 0;
         int distance = 0;
 
         if (previous != null) {
             previousCost = previous.cost;
-            distance = distanceBetween(previous.position, current);
-
-            boolean isTransport = distance > 1 || previous.position.getPlane() != current.getPlane();
+            distance = WorldPointUtil.distanceBetween(previous.packedPosition, packedPosition);
+            final int previousPlane = WorldPointUtil.unpackWorldPlane(previous.packedPosition);
+            final int currentPlane = WorldPointUtil.unpackWorldPlane(previous.packedPosition);
+            boolean isTransport = distance > 1 || previousPlane != currentPlane;
             if (isTransport) {
                 distance = wait;
             }
         }
 
         return previousCost + distance;
-    }
-
-    public static int distanceBetween(WorldPoint previous, WorldPoint current, int diagonal) {
-        int dx = Math.abs(previous.getX() - current.getX());
-        int dy = Math.abs(previous.getY() - current.getY());
-
-        if (diagonal == 1) {
-            return Math.max(dx, dy);
-        } else if (diagonal == 2) {
-            return dx + dy;
-        }
-
-        return Integer.MAX_VALUE;
-    }
-
-    public static int distanceBetween(WorldPoint previous, WorldPoint current) {
-        return distanceBetween(previous, current, 1);
     }
 }