… src/main/java/shortestpath/pathfinder/SplitFlagMap.java
… src/test/java/pathfinder/WorldPointTests.java