Yet another performance PR.

RuneLite's `WorldPoint` and `WorldArea` are immutable meaning any modifications to to them create an entirely new object. While that's good practice from an API design standpoint, it's not so good for performance-critical code like pathfinding. This PR packs all uses of `WorldPoint` in the hotpath into ints to prevent unnecessary allocations.

Changes made: 
* New `WorldPointUtil` with various helper functions to pack/unpack world points
  * Moved `Node.distanceBetween` here since it makes sense for that to live in a util file.
  * `WorldPointUtil.distanceToArea` should be guaranteed to match `WorldArea.distanceTo`. Added an additional test for clarity.
* Made a copy of `PathfinderConfig.transports` that uses the packed world point as a key. Doesn't copy the underlying lists so both `transports` and `transportsPacked` hold the same list references.
* `Pathfinder` no longer stores the entire path each time it finds a better one but instead stores the best last node in the path.
  * The path is lazily reconstructed and cached as needed rather than during the search.
* Removed instant and replaced with `System.currentTimeMillis()` to avoid making new `Instant` allocations.
* Decoupled `CollisionMap` from `SplitFlagMap`; `SplitFlagMap` is what stores data and is allocated once, `CollisionMap` holds a reference to `SplitFlagMap` and can be allocated once per thread.
  * Doing this allows me to remove the extra `boolean[]` and `ArrayList` allocations in `getNeighbors` without risk to thread-safety.
  * `ThreadLocal` incurs a hashmap lookup every time `get()` is called so I've also removed redundant calls to `plugin.getMap()`.

Main allocation cost now is making new nodes and a few-other more minor things ~~(can store a single boolean[] rather than [make a new one](https://github.com/Skretzo/shortest-path/blob/master/src/main/java/shortestpath/pathfinder/CollisionMap.java#L108) each time; same with the neighbors ArrayList)~~ (I've gone ahead and made this optimization, see first comment below). In theory, the node allocation cost can be addressed with an [Object Pool](https://en.wikipedia.org/wiki/Object_pool_pattern) (e.x. [Apache Commons' SoftReferenceObjectPool](https://commons.apache.org/proper/commons-pool/apidocs/org/apache/commons/pool2/impl/SoftReferenceObjectPool.html)), but it'd come with a not-insignificant runtime performance penalty.

# Results
Same setup as usual. See comment below for latest results with the `CollisionMap` changes.

**Worst-Case:** GE to top-left corner in the ocean
Didn't measure short-case this time as it doesn't make a huge difference for a small amount of nodes.

Forgive me for not including a table here, wanting to save some time. Overall this change ends up ~20ms faster with about 50% less allocated bytes (exact numbers in screenshots below).

## Before
**Worst-Case Time** (Avoid Wilderness): 156.9508ms
**Worst-Case Time**: 179.8608ms

**Memory Allocations:**
![image](https://github.com/Skretzo/shortest-path/assets/2230152/7e9453a2-2c7c-4945-81d0-74de6f08a2e1)

**Memory Allocations (Avoid Wilderness):**
![image](https://github.com/Skretzo/shortest-path/assets/2230152/2f30b667-219e-49c6-a95d-281155fa58e7)


## After
**Worst-Case Time** (Avoid Wilderness): 134.7586ms
**Worst-Case Time**: 156.9396ms

**Memory Allocations:**
<img width="607" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/97185fdb-6467-4f51-85aa-74730c1e0e28">

**Memory Allocations (Avoid Wilderness):**
<img width="494" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/efc6c16b-e929-4ca5-b8ba-ee688e0b332c">