@jocopa3 2023-07-17
**Update:** refactored `CollisionMap` and `SplitFlagMap` a bit. `CollisionMap` is now it's own thing while `SplitFlagMap` is what stores map data. This allows me to make `CollisionMap` thread-local which in-turn means the `boolean[]` and `ArrayList` allocations in `getNeighbors` can be made stateful while remaining thread-safe. This comes at no extra cost for performance while reducing allocations even further.

## Results

Worst-Case Time (Avoid Wilderness): **136.8438ms**
Worst-Case Time: **160.6458ms**

Note: I don't think this is actually any slower than the last commit, I think the few ms difference between this result and the previous commit comes down to different workloads running on my machine.

**Memory Allocations:**
<img width="554" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/b6704163-1125-4995-b76d-8ca60464126d">

**Memory Allocations (Avoid Wilderness):**
<img width="554" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/3d7653f0-fb28-4ca6-bd2b-2322f360e1cc">