@jocopa3 2023-07-17
Heads up: turns out the `isNear` check is redundant. It's meant to cancel pathfinding if the player isn't near the path, but it's already [checked in onGameTick](https://github.com/Skretzo/shortest-path/blob/master/src/main/java/shortestpath/ShortestPathPlugin.java#L190-L195) which calls `setTarget(null)` and [cancels pathfinding](https://github.com/Skretzo/shortest-path/blob/master/src/main/java/shortestpath/ShortestPathPlugin.java#L335) anyway. It seems it was only added to the pathfinder as there was no way to cancel pathfinding thread, but my previous threading change addressed that.

Allocations in the worst-case with Avoid Wilderness off now looks like:
<img width="578" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/6d1b8194-49e6-4a2a-a28f-2fa6b94abbb0">

I don't think additional optimization is needed after this. The only way to reduce this is to reduce the search space a la bidirectional BFS, but it's not super critical in my opinion. It's at a point where any other suggestions for optimization require making a tradeoff between runtime performance and memory allocation.