@jocopa3 2023-07-17
Addressed feedback and cleaned up some things that become unused after this PR.